Moved TG commands to a separate unit
