+++ src/service.ts
@@ -1,39 +1,8 @@
 import axios from "axios";
 import { Markup } from "telegraf";
-import { InlineKeyboardButton } from "typegram";
 import { CommunityResult } from "./api/types";
 import config from "./config";
 import logger from "./utils/logger";
-
-const helpCommand = (ctx: any): void => {
-  const helpHeader =
-    "Hello there! My name is Medousa.\n" +
-    "I'm part of the [Agora](https://agora-space.vercel.app/) project and " +
-    "I am your personal assistant.\n" +
-    "I will always let you know whether you can join a higher group or " +
-    "whether you were kicked from a group.\n";
-
-  let commandsList = "/help - show instructions\n/ping - check if I'm alive\n";
-
-  const helpFooter =
-    "For more details about me read the documentation on " +
-    "[github](https://github.com/AgoraSpaceDAO/telegram-runner).";
-
-  // DM
-  if (ctx.message.chat.id >= 0) {
-    commandsList +=
-      "/communities - get a list of your communities' websites\n" +
-      "/leave - you have to choose which community you want " +
-      "to leave and I'll do the rest\n";
-  } // group chat
-  else {
-    commandsList += "";
-  }
-
-  ctx.replyWithMarkdown(`${helpHeader}\n${commandsList}\n${helpFooter}`, {
-    disable_web_page_preview: true
-  });
-};
 
 const fetchCommunitiesOfUser = async (
   idFromPlatform: string
@@ -49,31 +18,6 @@ const leaveCommunity = (idFromPlatform: string, communityId: string): void => {
     })
     .then((res) => logger.debug(JSON.stringify(res.data)))
     .catch(logger.error);
-};
-
-const leaveCommand = (ctx: any): void => {
-  if (ctx.message.chat.id > 0) {
-    const communityList: InlineKeyboardButton[][] = [
-      [Markup.button.callback("Agora", "leave_confirm_0_Agora")]
-    ];
-
-    ctx.replyWithMarkdown(
-      "Choose the community you want to leave from the list below:",
-      Markup.inlineKeyboard(communityList)
-    );
-  }
-};
-
-const listCommunitiesCommand = (ctx: any): void => {
-  fetchCommunitiesOfUser(ctx.message.from.id).then((results) => {
-    const urls = results
-      .map((result) => `[${result.name}](${result.url})`)
-      .join("\n");
-
-    ctx.replyWithMarkdown(
-      `*Please visit your communities' websites:*\n${urls}`
-    );
-  });
 };
 
 const confirmLeaveCommunityAction = (ctx: any): void => {
@@ -98,11 +42,8 @@ const confirmedLeaveCommunityAction = (ctx: any): void => {
 };
 
 export {
-  helpCommand,
-  leaveCommand,
   leaveCommunity,
   fetchCommunitiesOfUser,
-  listCommunitiesCommand,
   confirmLeaveCommunityAction,
   confirmedLeaveCommunityAction
 };
